# backend/src/design/api/v2/openapi.yaml
openapi: 3.0.3
info:
  title: Learnify API
  version: "3.0"
  description: |
    Online Course Platform API — v3.
    Changes from v2:
    - Removed dead endpoints: /courses/featured, /courses/continue, /courses/stats, /lessons/:id/course, /payments/webhook-logs
    - Password validation: min 8 chars, 1 uppercase, 1 number
    - Email uniqueness check on profile update (409 if taken)
    - Pagination params clamped (page >= 1, limit 1-100)
    - Notifications response now includes pagination object
    - Course list: added enrollment, price, instructor_id filters
    - Webhook: improved auth validation, atomic payment update
    - Comments: lesson existence check, batched mention queries, content type validation
    - Bookmarks: course existence check (404)
    - Enrollments: progress object validation
    - Course delete: cascade deletes comments, mentions, notes, reviews, bookmarks
    - Upload: file type validation (MIME whitelist)
servers:
  - url: https://asm3-cos30043.vercel.app
    description: Production (Vercel)
  - url: http://localhost:3000
    description: Local dev
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Payment:
      type: object
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        course_id: { type: string, format: uuid }
        amount: { type: number }
        method: { type: string }
        transaction_id: { type: string }
        status: { type: string, enum: [pending, completed, failed] }
        created_at: { type: string, format: date-time }
    Review:
      type: object
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        course_id: { type: string, format: uuid }
        rating: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string, nullable: true }
        username: { type: string }
        avatar_url: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    Note:
      type: object
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        lesson_id: { type: string, format: uuid }
        content: { type: string }
        timestamp_sec: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    Pagination:
      type: object
      description: "page >= 1, limit clamped to 1-100"
      properties:
        page: { type: integer, minimum: 1 }
        limit: { type: integer, minimum: 1, maximum: 100 }
        total: { type: integer }
        pages: { type: integer }
...
paths:
...
  /api/payments/create:
    post:
      tags: [Payments]
      summary: Create payment (SePay QR)
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [course_id]
              properties:
                course_id: { type: string, format: uuid }
                method: { type: string, default: sepay }
      responses:
        "201":
          description: Payment created with QR info
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: object
                    properties:
                      payment: { $ref: "#/components/schemas/Payment" }
                      qr_info:
                        type: object
                        properties:
                          amount: { type: number }
                          transaction_id: { type: string }
                          description: { type: string }
  /api/payments/{id}/status:
    get:
      tags: [Payments]
      summary: Check payment status
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Payment status
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: "#/components/schemas/Payment" }
  /api/payments/webhook:
    post:
      tags: [Payments]
      summary: SePay webhook callback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [transaction_id, status, api_key]
              properties:
                transaction_id: { type: string }
                status: { type: string }
                api_key: { type: string }
      responses:
        "200": { description: Processed }
        "401": { description: Invalid API key }

  # ── Reviews ──────────────────────────────────────────────
  /api/courses/{courseId}/reviews:
    get:
      tags: [Reviews]
      summary: List reviews with aggregate ratings
      parameters:
        - name: courseId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
      responses:
        "200":
          description: Reviews with aggregate
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: object
                    properties:
                      reviews:
                        type: array
                        items: { $ref: "#/components/schemas/Review" }
                      user_review: { $ref: "#/components/schemas/Review" }
                      aggregate:
                        type: object
                        properties:
                          avg_rating: { type: number }
                          total_reviews: { type: integer }
                          distribution:
                            type: object
                            properties:
                              "1": { type: integer }
                              "2": { type: integer }
                              "3": { type: integer }
                              "4": { type: integer }
                              "5": { type: integer }
                      pagination: { $ref: "#/components/schemas/Pagination" }

    post:
      tags: [Reviews]
      summary: Create review (must be enrolled, one per user per course)
      security: [{ BearerAuth: [] }]
      parameters:
        - name: courseId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rating]
              properties:
                rating: { type: integer, minimum: 1, maximum: 5 }
                comment: { type: string }
      responses:
        "201": { description: Review created }
        "400": { description: Invalid rating }
        "403": { description: Not enrolled }
        "409": { description: Already reviewed }
  /api/reviews/{id}:
    put:
      tags: [Reviews]
      summary: Update own review
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                rating: { type: integer, minimum: 1, maximum: 5 }
                comment: { type: string }
      responses:
        "200": { description: Updated }
        "403": { description: Not owner }
        "404": { description: Not found }

    delete:
      tags: [Reviews]
      summary: Delete own review
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: Deleted }
        "403": { description: Not owner }
        "404": { description: Not found }

  # ── Bookmarks ────────────────────────────────────────────
...
  /api/lessons/{lessonId}/notes:
    get:
      tags: [Notes]
      summary: List notes for a lesson (own notes only)
      security: [{ BearerAuth: [] }]
      parameters:
        - name: lessonId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Notes sorted by timestamp
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Note" }

    post:
      tags: [Notes]
      summary: Create timestamped note
      security: [{ BearerAuth: [] }]
      parameters:
        - name: lessonId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string }
                timestamp_sec: { type: integer, default: 0, description: "Video timestamp in seconds" }
      responses:
        "201": { description: Note created }
        "400": { description: Content required }
  /api/notes/{id}:
    put:
      tags: [Notes]
      summary: Update own note
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string }
      responses:
        "200": { description: Updated }
        "403": { description: Not owner }
        "404": { description: Not found }

    delete:
      tags: [Notes]
      summary: Delete own note
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: Deleted }
        "403": { description: Not owner }
        "404": { description: Not found }

  # ── Notifications ────────────────────────────────────────
...